refactor: move fee entity schemas with their inferred types
