+++ src/modules/fees/domain/entities/__tests__/can-relay-response.entity.spec.ts
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: FSL-1.1-MIT
-import { CanRelayResponseSchema } from '@/modules/fees/domain/entities/schemas/can-relay-response.schema';
+import { CanRelayResponseSchema } from '@/modules/fees/domain/entities/can-relay-response.entity';
 
 describe('CanRelayResponseSchema', () => {
   it('should validate a valid can-relay response', () => {
