+++ src/modules/fees/domain/entities/__tests__/gtf-fees-request.entity.spec.ts
@@ -2,8 +2,8 @@
 
 import { faker } from '@faker-js/faker';
 import { getAddress, zeroAddress } from 'viem';
+import { GtfFeesRequestSchema } from '@/modules/fees/domain/entities/gtf-fees-request.entity';
 import { Origin } from '@/modules/fees/domain/entities/origin.entity';
-import { GtfFeesRequestSchema } from '@/modules/fees/domain/entities/schemas/gtf-fees-request.schema';
 
 describe('GtfFeesRequestSchema', () => {
   it('should validate a valid gtf-fees request', () => {
