+++ src/modules/fees/domain/entities/__tests__/gtf-fees-response.entity.spec.ts
@@ -3,14 +3,14 @@
 import { faker } from '@faker-js/faker';
 import { getAddress, zeroAddress } from 'viem';
 import { gtfFeesResponseBuilder } from '@/modules/fees/domain/entities/__tests__/gtf-fees-response.builder';
-import { Origin } from '@/modules/fees/domain/entities/origin.entity';
 import {
   GtfFeeBreakdownSchema,
   GtfFeesResponseSchema,
   GtfPricingContextSnapshotSchema,
   GtfTxDataSchema,
   GtfValuationDetailSchema,
-} from '@/modules/fees/domain/entities/schemas/gtf-fees-response.schema';
+} from '@/modules/fees/domain/entities/gtf-fees-response.entity';
+import { Origin } from '@/modules/fees/domain/entities/origin.entity';
 
 describe('GtfTxDataSchema', () => {
   const validTxData = {
