--- src/modules/fees/domain/entities/schemas/can-relay-response.schema.ts
@@ -1,6 +1,0 @@
-// SPDX-License-Identifier: FSL-1.1-MIT
-import { z } from 'zod';
-
-export const CanRelayResponseSchema = z.object({
-  canRelay: z.boolean(),
-});
