--- src/modules/fees/domain/entities/schemas/gas-token.schema.ts
@@ -1,11 +1,0 @@
-// SPDX-License-Identifier: FSL-1.1-MIT
-import { z } from 'zod';
-import { buildLenientPageSchema } from '@/domain/entities/schemas/page.schema.factory';
-import { AddressSchema } from '@/validation/entities/schemas/address.schema';
-
-export const GasTokenSchema = z.object({
-  address: AddressSchema,
-  symbol: z.string(),
-});
-
-export const GasTokenLenientPageSchema = buildLenientPageSchema(GasTokenSchema);
