+++ src/modules/fees/domain/entities/tx-fees-request.entity.ts
@@ -1,6 +1,22 @@
 // SPDX-License-Identifier: FSL-1.1-MIT
-
-import type { z } from 'zod';
-import type { TxFeesRequestSchema } from '@/modules/fees/domain/entities/schemas/tx-fees-request.schema';
+import { z } from 'zod';
+import { Origin } from '@/modules/fees/domain/entities/origin.entity';
+import { Operation } from '@/modules/safe/domain/entities/operation.entity';
+import { AddressSchema } from '@/validation/entities/schemas/address.schema';
+import { HexSchema } from '@/validation/entities/schemas/hex.schema';
+import { NonNegativeNumericStringSchema } from '@/validation/entities/schemas/non-negative-numeric-string.schema';
+import { NumericStringSchema } from '@/validation/entities/schemas/numeric-string.schema';
 
 export type TxFeesRequest = z.infer<typeof TxFeesRequestSchema>;
+
+export const TxFeesRequestSchema = z.object({
+  to: AddressSchema,
+  value: NumericStringSchema,
+  data: HexSchema,
+  operation: z.enum(Operation),
+  numberSignatures: z.number().int().min(1),
+  nonce: NonNegativeNumericStringSchema,
+  gasToken: AddressSchema,
+  origin: z.enum(Origin).optional(),
+  fiatCode: z.string().optional(),
+});
